move ProductCarrousel to ProductPage folder, correct props
